# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: NyelvSzó API
  description: |
    Angol-magyar Nyelvészeti Szakszótár API dokumentáció.
    
    Ez az API lehetővé teszi a nyelvészeti szakkifejezések keresését, létrehozását, módosítását és törlését.
    Három felhasználói szerepkör létezik:
    - **Felhasználó (1)**: Csak megtekintés
    - **Szerkesztő (2)**: Szócikkek kezelése
    - **Adminisztrátor (3)**: Teljes hozzáférés
    
    ## Autentikáció
    Az API JWT (JSON Web Token) alapú autentikációt használ. A token-t a `/login` végponton lehet megszerezni.
    
    ## Rate Limiting
    - Általános végpontok: 100 kérés / 15 perc / IP
    - Login végpont: 5 kérés / 15 perc / IP
    
  version: 2.2.0
  contact:
    name: Dr. Ádám Porkoláb
    email: adam@porkolab.digital
    url: https://nyelvszo.eu
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.nyelvszo.eu
    description: Production server
paths:
  /login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive JWT access token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "user@example.com"
              password: "SecurePassword123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /login/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
              required:
                - refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  accessToken:
                    type: string
                  expiresIn:
                    type: string
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication. Add 'Bearer ' prefix to your JWT token.
        Example: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          format: objectId
          description: Unique identifier
          example: "507f1f77bcf86cd799439011"
        firstName:
          type: string
          description: First name
          example: "János"
        lastName:
          type: string
          description: Last name
          example: "Kovács"
        email:
          type: string
          format: email
          description: Email address
          example: "janos.kovacs@example.com"
        role:
          type: integer
          description: User role (1=User, 2=Editor, 3=Admin)
          enum: [1, 2, 3]
          example: 2
        isActive:
          type: boolean
          description: Whether user is active
          example: true
        lastLogin:
          type: string
          format: date-time
          description: Last login timestamp
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - _id
        - firstName
        - lastName
        - email
        - role
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        password:
          type: string
          description: User password
          example: "SecurePassword123!"
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Login success status
          example: true
        accessToken:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
        expiresIn:
          type: string
          description: Token expiration time
          example: "1h"
      required:
        - success
        - accessToken
        - user
        - expiresIn
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid credentials"
        message:
          type: string
          description: Detailed error message
          example: "The provided email or password is incorrect"
        statusCode:
          type: integer
          description: HTTP status code
          example: 401
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        path:
          type: string
          description: Request path where error occurred
          example: "/login"
      required:
        - error
        - statusCode
...